fix issue with fuzzy search
